+++ src/Sql/Sql.Test/ScenarioTests/DatabaseReplicationTests.ps1
@@ -350,7 +350,6 @@ function Create-DatabaseForTest  ($rg, $server, $edition = "Premium")
 	New-AzSqlDatabase -ResourceGroupName $rg.ResourceGroupName -ServerName $server.ServerName -DatabaseName $databaseName -Edition $edition
 }
 
-
 <#
 	.SYNOPSIS
 	Creates test database
@@ -359,4 +358,69 @@ function Create-VcoreDatabaseForTest  ($rg, $server, $numCores = 2, $licenseType
 {
 	$databaseName = Get-DatabaseName
 	New-AzSqlDatabase -ResourceGroupName $rg.ResourceGroupName -ServerName $server.ServerName -DatabaseName $databaseName -VCore $numCores -ComputeGeneration Gen5 -Edition GeneralPurpose -LicenseType $licenseType
+}
+
+<#
+	.SYNOPSIS
+	Creates test database with BackupStorageRedundancy
+#>
+function Test-CreateDatabaseCopyWithBackupStorageRedundancy()
+{
+	# Setup
+	$location = "eastus"
+	$rg = Create-ResourceGroupForTest $location
+	$server = Create-ServerForTest $rg $location
+	$database = Create-DatabaseForTest $rg $server "Standard"
+
+	$copyRg = Create-ResourceGroupForTest $location
+	$copyServer = Create-ServerForTest $copyRg $location
+	$copyDatabaseName = Get-DatabaseName
+
+	try
+	{
+		# Create a local database copy
+		$dbLocalCopy = New-AzSqlDatabaseCopy -ResourceGroupName $rg.ResourceGroupName -ServerName $server.ServerName -DatabaseName $database.DatabaseName `
+		 -CopyDatabaseName $copyDatabaseName -BackupStorageRedundancy zone
+
+		$newDb = Get-AzSqlDatabase -ResourceGroupName $dbLocalCopy.ResourceGroupName -ServerName $dbLocalCopy.ServerName -DatabaseName $copyDatabaseName
+		Assert-AreEqual "Zone" $newDb.BackupStorageRedundancy 
+
+		# Create a cross server copy
+	}
+	finally
+	{
+		Remove-ResourceGroupForTest $rg
+		Remove-ResourceGroupForTest $copyRg
+	}
+}
+
+<#
+	.SYNOPSIS
+	Tests creating a secondary database
+#>
+function Test-CreateSecondaryDatabaseWithBackupStorageRedundancy()
+{
+	# Setup
+    $location = "westcentralus"
+	$rg = Create-ResourceGroupForTest $location
+	$server = Create-ServerForTest $rg $location
+	$database = Create-DatabaseForTest $rg $server
+
+	$partRg = Create-ResourceGroupForTest $location
+	$partServer = Create-ServerForTest $partRg $location
+
+	try
+	{
+		# Create Readable Secondary
+		$readSecondary = New-AzSqlDatabaseSecondary -ResourceGroupName $rg.ResourceGroupName -ServerName $server.ServerName -DatabaseName $database.DatabaseName `
+		 -PartnerResourceGroupName $partRg.ResourceGroupName -PartnerServerName $partServer.ServerName -AllowConnections All -BackupStorageRedundancy local
+
+		$secondaryDb = Get-AzSqlDatabase -ResourceGroupName $readSecondary.PartnerResourceGroupName -ServerName $readSecondary.PartnerServerName -DatabaseName $readSecondary.DatabaseName
+		Assert-AreEqual $secondaryDb.BackupStorageRedundancy "Local"
+	}
+	finally
+	{
+		Remove-ResourceGroupForTest $rg
+		Remove-ResourceGroupForTest $partRg
+	}
 }